scrapyjob/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
pandas
numpy
scrapy
itemadapter
pymongo
mysql.connector
psycopg2